Second example playground
This playground demonstrates a simple UI: the user can tap a button to change the background color of the view. This one didn't get shown during the talk, but it's a nice, simple playground that demonstrates the techniques. Remember to type command-option-return to show the live view of the UI.